+++ CHANGELOG.md
@@ -35,6 +35,15 @@
   * Make `API_RESOURCE_KEY` in `LocalPaymentNonce` package-private
   * Make `LocalPaymentNonce#fromJson()` methods package-private
   * Make `LocalPaymentNonce` constructor package-private
+  * The `shippingMethod` property on `ThreeDSecureRequest` is now an enum rather than a string. Possible values:
+    * `SAME_DAY`
+    * `EXPEDITED`
+    * `PRIORITY`
+    * `GROUND`
+    * `ELECTRONIC_DELIVERY`
+    * `SHIP_TO_STORE`
+  * Change default `versionRequested` on `ThreeDSecureRequest` to `ThreeDSecureVersion.VERSION_2`
+  * Rename `uiCustomization` on `ThreeDSecureRequest` to `v2UiCustomization`
 
 ## 4.0.0-beta1